tests/wpt/tests/html/canvas/offscreen/the-canvas-state/2d.state.saverestore.fillStyle.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.state.saverestore.fillStyle</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.state.saverestore.fillStyle</h1>
<p class="desc">save()/restore() works for fillStyle</p>


<script>
var t = async_test("save()/restore() works for fillStyle");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  // Test that restore() undoes any modifications
  var old = ctx.fillStyle;
  ctx.save();
  ctx.fillStyle = "#ff0000";
  ctx.restore();
  _assertSame(ctx.fillStyle, old, "ctx.fillStyle", "old");

  // Also test that save() doesn't modify the values
  ctx.fillStyle = "#ff0000";
  old = ctx.fillStyle;
      // we're not interested in failures caused by get(set(x)) != x (e.g.
      // from rounding), so compare against 'old' instead of against "#ff0000"
  ctx.save();
  _assertSame(ctx.fillStyle, old, "ctx.fillStyle", "old");
  ctx.restore();
  t.done();

});
</script>
